requirements.txt
streamlit
openpyxl
pandas
requests
pillow
numpy
opencv-python-headless
urllib3
